[#17] Enhance Redis monitoring with command counter
Add RedisCommandCounter middleware for tracking Redis operations. This
complements the existing RedisLogger, providing more tools for
performance monitoring and debugging. The counter is thread-safe and can
be easily enabled or disabled as needed.
